broke out noteRoutes into own file
